move RotationState into it's own class
